Prevent theme switching with no delay
I was seeing an issue right along the boundary of sunrise/sunset where
it would constantly notify that the theme had been changed with an
extremely short delay and would continue like this until I had restart
nvim. While investigating this, I also noticed that the timers are in
ms, but this was being calculated in seconds.

I converted the vim.loop timer to vim.defer_fn. vim.loop as been
deprecated in favor of vim.uv anyways, and vim.defer_fn is more
aligned with what is trying to be done here (one-off timers being
created at varying frequency).